app/requirements.txt
fastapi
uvicorn[standard]
weaviate-client>=3.26.7
fpdf
requests
pdfplumber
pandas
python-docx
python-pptx>=0.6.21
textract